+++ examples/data-sources/meraki_appliance_ssids/data-source.tf
@@ -1,0 +1,3 @@
+data "meraki_appliance_ssids" "example" {
+  network_id = "L_123456"
+}